+++ src/test/java/org/folio/mr/repository/MediatedBatchRequestSplitRepositoryIT.java
@@ -12,8 +12,8 @@ import java.util.UUID;
 import org.folio.mr.api.BaseIT;
 import org.folio.mr.domain.BatchRequestSplitStatus;
 import org.folio.mr.domain.BatchRequestStatus;
-import org.folio.mr.domain.entity.BatchRequest;
-import org.folio.mr.domain.entity.BatchRequestSplit;
+import org.folio.mr.domain.entity.MediatedBatchRequest;
+import org.folio.mr.domain.entity.MediatedBatchRequestSplit;
 import org.folio.mr.domain.entity.MetadataEntity;
 import org.folio.spring.FolioExecutionContext;
 import org.folio.test.types.IntegrationTest;
@@ -22,31 +22,31 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.bean.override.mockito.MockitoSpyBean;
 
 @IntegrationTest
-class BatchRequestSplitRepositoryIT extends BaseIT {
+class MediatedBatchRequestSplitRepositoryIT extends BaseIT {
 
   @MockitoSpyBean
   private FolioExecutionContext context;
 
   @Autowired
-  private BatchRequestRepository batchRequestRepository;
+  private MediatedBatchRequestRepository batchRequestRepository;
 
   @Autowired
-  private BatchRequestSplitRepository batchRequestSplitRepository;
+  private MediatedBatchRequestSplitRepository batchRequestSplitRepository;
 
   @Test
   void batchRequestSplitMetadataShouldBeGeneratedWhenSaveEntity() {
     var randomUUID = UUID.randomUUID();
     var userId = UUID.randomUUID();
     var requestDate = Timestamp.from(Instant.parse("2025-09-10T10:15:30.00Z"));
-    var batch = BatchRequest.builder()
+    var batch = MediatedBatchRequest.builder()
       .id(randomUUID)
       .requesterId(randomUUID)
       .status(BatchRequestStatus.fromValue("In progress"))
       .requestDate(requestDate)
       .build();
-    var requestSplit = BatchRequestSplit.builder()
+    var requestSplit = MediatedBatchRequestSplit.builder()
       .id(randomUUID)
-      .batchRequest(batch)
+      .mediatedBatchRequest(batch)
       .requesterId(randomUUID)
       .status(BatchRequestSplitStatus.fromValue("Completed"))
       .build();
@@ -65,11 +65,11 @@ class BatchRequestSplitRepositoryIT extends BaseIT {
       .containsOnly(userId, userId);
 
     assertThat(savedBatch)
-      .extracting(BatchRequest::getRequesterId, BatchRequest::getRequestDate, BatchRequest::getStatus)
+      .extracting(MediatedBatchRequest::getRequesterId, MediatedBatchRequest::getRequestDate, MediatedBatchRequest::getStatus)
       .containsExactly(randomUUID, requestDate, IN_PROGRESS);
 
     assertThat(savedSplit)
-      .extracting(BatchRequestSplit::getRequesterId, BatchRequestSplit::getStatus)
+      .extracting(MediatedBatchRequestSplit::getRequesterId, MediatedBatchRequestSplit::getStatus)
       .containsExactly(randomUUID, COMPLETED);
   }
 }