+++ NEWS.md
@@ -1,5 +1,6 @@
 ## 1.3.0-SNAPSHOT 2025-XX-XX
 * Add Batch Request DB tables ([MODREQMED-159](https://folio-org.atlassian.net/browse/MODREQMED-159))
+* Add APIs for Mediated Batch Request and Request Details ([MODREQMED-159](https://folio-org.atlassian.net/browse/MODREQMED-158))
 
 ## 1.2.0 2025-03-13
 * Upgrade to Java 21, folio-spring-support 9.0.0 (MODREQMED-97)